requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
python-telegram-bot==21.1.1
python-telegram-bot[job-queue]